requirements.txt
Jinja2==2.10
openpyxl==2.5.3
paramiko==2.4.1
requests==2.18.4
PyYAML==5.3.1
regex